[breaking] Moved template code to headers
