+++ rdmalib/include/rdmalib/rdmalib.hpp
@@ -8,6 +8,7 @@
 #include <vector>
 #include <mutex>
 #include <functional>
+#include <cereal/archives/json.hpp>
 
 // #ifdef USE_LIBFABRIC
 #include <rdma/fabric.h>
@@ -23,6 +24,7 @@ extern "C" {
 
 #include <rdmalib/buffer.hpp>
 #include <rdmalib/connection.hpp>
+#include <rdmalib/server.hpp>
 #include <rdmalib/libraries.hpp>
 
 namespace rdmalib {
@@ -237,6 +239,38 @@ namespace rdmalib {
     void accept(VerbsConnection* connection);
     void set_nonblocking_poll();
   };
+
+  namespace server {
+
+    template <typename Library>
+    ServerStatus<Library>::ServerStatus():
+      _address(""),
+      _port(0)
+    {}
+
+    template <typename Library>
+    ServerStatus<Library>::ServerStatus(std::string address, int port):
+      _address(address),
+      _port(port)
+    {}
+
+    template <typename Library>
+    ServerStatus<Library> ServerStatus<Library>::deserialize(std::istream & in)
+    {
+      ServerStatus status;
+      cereal::JSONInputArchive archive_in(in);
+      archive_in(status);
+      return status;
+    }
+
+    template <typename Library>
+    void ServerStatus<Library>::serialize(std::ostream & out) const
+    {
+      cereal::JSONOutputArchive archive_out(out);
+      archive_out(*this);
+    }
+  }
+
 }
 
 #endif
