+++ rfaas/include/rfaas/executor.hpp
@@ -698,6 +698,42 @@ namespace rfaas {
     }
   };
 
+  template <typename Library>
+  executor_state<Library>::executor_state(Connection_t* conn, int rcv_buf_size):
+    conn(conn),
+    _rcv_buffer(rcv_buf_size)
+  {
+  }
+
+  template <typename Derived, typename Library>
+  executor<Derived, Library>::executor(std::string address, int port, int rcv_buf_size, int max_inlined_msg):
+    _state(address, port, rcv_buf_size + 1),
+    _rcv_buffer(rcv_buf_size),
+    _execs_buf(MAX_REMOTE_WORKERS),
+    _address(address),
+    _port(port),
+    _rcv_buf_size(rcv_buf_size),
+    _executions(0),
+    _max_inlined_msg(max_inlined_msg),
+    _perf(1000)
+  {
+    events = 0;
+    _active_polling = false;
+    _end_requested = false;
+  }
+
+  template <typename Derived, typename Library>
+  executor<Derived, Library>::executor(device_data & dev):
+    executor(dev.ip_address, dev.port, dev.default_receive_buffer_size, dev.max_inline_data)
+  {}
+
+  template <typename Derived, typename Library>
+  executor<Derived, Library>::~executor()
+  {
+    this->deallocate();
+    _perf.export_csv("client_perf.csv", {"start", "function parsed", "function post written", "buffer refilled", "received result", "parsed result", "catched unlikely case", "polled send"});
+  }
+
 }
 
 #endif
