+++ server/executor/opts.hpp
@@ -1,7 +1,7 @@
 
 #include <cxxopts.hpp>
 
-#include "server.hpp"
+// #include "server.hpp"
 
 namespace server {
   namespace types {
